Moved auto sale category assignment into its own class.
