+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.30.11",
+  "version": "1.30.12",
   "title": "shop-standards",
   "description": "Standard refinements for e-commerce websites.",
   "dependencies": {
