+++ src/Plugin.php
@@ -102,11 +102,7 @@ class Plugin {
     // Adds woocommerce specific settings.
     add_filter('woocommerce_get_settings_shop_standards', __NAMESPACE__ . '\WooCommerce::woocommerce_get_settings_shop_standards');
 
-    // Assigns sale category conditionally on product update.
-    if (get_option('_' . Plugin::L10N . '_enable_auto_sale_category_assignment') === 'yes') {
-      add_action('woocommerce_update_product', __NAMESPACE__ . '\WooCommerce::woocommerce_update_product');
-    }
-
+    WooCommerceSaleLabel::init();
     Seo::init();
     WooCommerceSalutation::init();
